+++ packages/calcite-components/src/components/input-time-zone/input-time-zone.e2e.ts
@@ -11,39 +11,70 @@ import {
   reflects,
   renders,
   t9n,
+  TagAndPage,
 } from "../../tests/commonTests";
 import { toUserFriendlyName } from "./utils";
 
 describe("calcite-input-time-zone", () => {
-  describe.skip("accessible", () => {
-    accessible(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
+  const testTimeZoneNamesAndOffsets = [
+    { name: "America/Los_Angeles", offset: -420, label: "GMT-7" },
+    { name: "America/Denver", offset: -360, label: "GMT-6" },
+    { name: "Europe/London", offset: 60, label: "GMT+1" },
+  ];
+
+  async function simpleTestProvider(): Promise<TagAndPage> {
+    const page = await newE2EPage();
+    await page.emulateTimezone(testTimeZoneNamesAndOffsets[0].name);
+    await page.setContent(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
+
+    return {
+      page,
+      tag: "calcite-input-time-zone",
+    };
+  }
+
+  describe("accessible", () => {
+    accessible(simpleTestProvider);
   });
 
-  describe.skip("focusable", () => {
-    focusable(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
+  describe("focusable", () => {
+    focusable(simpleTestProvider);
   });
 
-  describe.skip("formAssociated", () => {
-    formAssociated(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`), {
-      testValue: "-360",
-      clearable: false,
+  describe("formAssociated", () => {
+    formAssociated(
+      {
+        tagOrHTML: addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`),
+        beforeContent: async (page) => {
+          await page.emulateTimezone(testTimeZoneNamesAndOffsets[0].name);
+        },
+      },
+      {
+        testValue: "-360",
+        clearable: false,
+      }
+    );
+  });
+
+  describe("hidden", () => {
+    hidden(simpleTestProvider);
+  });
+
+  describe("renders", () => {
+    renders(simpleTestProvider, { display: "block" });
+  });
+
+  describe("labelable", () => {
+    labelable({
+      tagOrHTML: addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`),
+      beforeContent: async (page) => {
+        await page.emulateTimezone(testTimeZoneNamesAndOffsets[0].name);
+      },
     });
   });
 
-  describe.skip("hidden", () => {
-    hidden(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
-  });
-
-  describe.skip("renders", () => {
-    renders(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`), { display: "block" });
-  });
-
-  describe.skip("labelable", () => {
-    labelable(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
-  });
-
-  describe.skip("reflects", () => {
-    reflects(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`), [
+  describe("reflects", () => {
+    reflects(simpleTestProvider, [
       { propertyName: "disabled", value: true },
       { propertyName: "maxItems", value: 0 },
       { propertyName: "mode", value: "offset" },
@@ -53,8 +84,8 @@ describe("calcite-input-time-zone", () => {
     ]);
   });
 
-  describe.skip("defaults", () => {
-    defaults(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`), [
+  describe("defaults", () => {
+    defaults(simpleTestProvider, [
       { propertyName: "disabled", defaultValue: false },
       { propertyName: "maxItems", defaultValue: 0 },
       { propertyName: "messageOverrides", defaultValue: undefined },
@@ -65,21 +96,15 @@ describe("calcite-input-time-zone", () => {
     ]);
   });
 
-  describe.skip("disabled", () => {
-    disabled(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`), {
+  describe("disabled", () => {
+    disabled(simpleTestProvider, {
       shadowAriaAttributeTargetSelector: "calcite-combobox",
     });
   });
 
-  describe.skip("t9n", () => {
-    t9n(addTimeZoneNamePolyfill(html` <calcite-input-time-zone></calcite-input-time-zone>`));
+  describe("t9n", () => {
+    t9n(simpleTestProvider);
   });
-
-  const testTimeZoneNamesAndOffsets = [
-    { name: "America/Los_Angeles", offset: -420, label: "GMT-7" },
-    { name: "America/Denver", offset: -360, label: "GMT-6" },
-    { name: "Europe/London", offset: 60, label: "GMT+1" },
-  ];
 
   describe("mode", () => {
     describe("offset (default)", () => {
@@ -133,6 +158,22 @@ describe("calcite-input-time-zone", () => {
         const timeZoneItem = await page.find("calcite-input-time-zone >>> calcite-combobox-item[selected]");
 
         expect(await timeZoneItem.getProperty("textLabel")).toMatch(testTimeZoneNamesAndOffsets[0].label);
+      });
+
+      it("omits filtered or non-localized time zones (incoming to browser)", async () => {
+        const page = await newE2EPage();
+        await page.emulateTimezone(testTimeZoneNamesAndOffsets[0].name);
+        await page.setContent(
+          await addTimeZoneNamePolyfill(html` <calcite-input-time-zone value="60"></calcite-input-time-zone>`)
+        );
+
+        const input = await page.find("calcite-input-time-zone");
+
+        expect(await input.getProperty("value")).toBe(`${testTimeZoneNamesAndOffsets[2].offset}`);
+
+        const timeZoneItem = await page.find("calcite-input-time-zone >>> calcite-combobox-item[selected]");
+
+        expect(await timeZoneItem.getProperty("textLabel")).toMatch(testTimeZoneNamesAndOffsets[2].label);
       });
     });
 
@@ -244,7 +285,7 @@ describe("calcite-input-time-zone", () => {
  * @param testHtml
  */
 function addTimeZoneNamePolyfill(testHtml: string): string {
-  return html` <script>
+  return html` <script type="module">
       const OriginalDateTimeFormat = Intl.DateTimeFormat;
 
       class ExtendedDateTimeFormat extends OriginalDateTimeFormat {
@@ -269,7 +310,7 @@ function addTimeZoneNamePolyfill(testHtml: string): string {
                 ? "-7"
                 : timeZone === "America/Denver"
                 ? "-6"
-                : timeZone === "Europe/London"
+                : timeZone === "Europe/London" || timeZone === "Europe/Belfast" || timeZone === "Etc/GMT-1"
                 ? "+1"
                 : "+0");
 
@@ -298,7 +339,15 @@ function addTimeZoneNamePolyfill(testHtml: string): string {
 
       Intl.supportedValuesOf = function (key) {
         if (key === "timeZone") {
-          return ["America/Los_Angeles", "America/Denver", "Europe/London"];
+          return [
+            "America/Los_Angeles",
+            "America/Denver",
+            "Europe/London",
+
+            // not available in Chromium v92 at time of testing
+            "Etc/GMT-1",
+            "Europe/Belfast",
+          ];
         }
       };
     </script>